requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
python-dotenv==1.0.1
gunicorn==21.2.0
